requirements.txt
blueness
blue-options
abcli
openai_commands
blueflow

boto3
openai
geojson
geopandas
matplotlib
numpy
opencv-python
pymysql
pyyaml
pylint
pytest
python-dotenv[cli]
tqdm
beautifulsoup4

networkx
pydot

# for RAG
transformers
torch
faiss-cpu